Update README.md with necessary stack.
Update README.md with the necessary stack used in the project development. 
This will help users to EASILY visualize the languages used in the entire project without trying to figure that out per repository. 
Also will help contributors easily visualize which languages are needed and can be used to further support the project.