Update SECURITY.md
# UPDATED.policy
You may also reach out to the owner will drive it through the fix and release process.Work with the organisation to determine if this security report requires if an advisory is required. 
One must be the repo admin to do this. who are not a repo admin will reach out to the creatoruntil they find who can create the advisory. The repo admins who are also owner of this advisory and repositories @kingabnt00 @abent999
parmenantly closed [Issue 72555](https://github.com/flutter/flutter/issues/72555)